# tasks/security/sec-009-api-security-audit/environment/data/api_spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "User Management API",
    "version": "2.1.0"
  },
  "servers": [
    {"url": "https://api.example.com/v2"}
  ],
  "security": [],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "username": {"type": "string"},
          "email": {"type": "string"},
          "ssn": {"type": "string", "description": "Social Security Number"},
          "phone": {"type": "string"},
          "date_of_birth": {"type": "string", "format": "date"},
          "address": {"type": "string"},
          "credit_card_last4": {"type": "string"}
        }
      }
    }
  },
...
  "paths": {
    "/users": {
      "get": {
        "summary": "List all users with full details",
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of user objects including SSN, DOB, address",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {"$ref": "#/components/schemas/User"}
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {"type": "string"},
                  "email": {"type": "string"},
                  "password": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "User created"}
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get user by ID",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "User object",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/User"}
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a user",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "204": {"description": "User deleted"}
        }
      }
    },
...
  }
}